+++ tests/fixtures/source/MainWhile.xs
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+fn main() -> Long {
+  value: Long = 3;
+  while (value > 0) {
+    value -= 1;
+  }
+  return value + 7;
+}